+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -79,7 +79,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-24T04:50:03.169Z")
+    created: Optional[datetime] = parse_datetime("2025-10-24T08:57:05.097Z")
     r"""The date and time the resource was created"""